cleaning non test its compiling
